refactor: nav tabs, routing, layout, home
