+++ src/routes.js
@@ -1,8 +1,6 @@
 const routes = {
   _: '/',
-  pro: '/pro',
   about: '/about',
-  social: '/social',
 };
 
 export default routes;
